Updated examples/http_proxy.rb
The current example was not compatible with Ruby 1.8.7 due to use of
the 'String#clear' method. Also, I added a note to let the user know
that if they would like to use this example as a starting point, they
should remove all the 'puts' statements for performance and also to
avoid the return of corrupt images(as is such in the case of loading
maps.google.com through the proxy).